refactor(tools/risk-form-filler): split up classes into component files
