+++ apps/docs/content/docs/contribution-guidelines/coding-standards/tooling.mdx
@@ -20,6 +20,8 @@ Originally authored by [Jeremy Friesen](https://github.com/JeremyFriesenGitHub).
   Your browser does not support the video tag.
 </video>
 
+Configurations can be found in `config.data.ts`.
+
 ## Sponsorship
 ### Email Drafts
